style: dọn 6 chỗ thừa, không đổi hành vi
- `rhai_collect.rs`: bỏ `use inventory;` thừa (đường dẫn đã nằm trong prelude
  từ edition 2018).
- `ts_ops.rs`: bỏ hai ép kiểu `x as i64` trên biến đã là `i64`
  (`ts_moving_avg`, `ts_resample`).
- `rhai_func.rs` / `rhai_register.rs`: bỏ `mut` không dùng.
- `transition.rs`: `t.grid().clone()` → `*t.grid()` (`AnalysisGrid` là `Copy`
  nên clone thừa một lần).
- `test_predict_strategy.rs`: `expect(&format!(...))` →
  `unwrap_or_else(|| panic!(...))`. `expect` nhận `&str` nên `&format!()` ép
  `String` rồi `Deref` — tức **format dù có thành công hay không** mỗi lần gọi
  `find` (clippy `expect_fun_call`).

Build sạch. `opsense-rhai` 54/54, `opsense-macros` 13/13, `opsense-mlib` 105/105.

+++ crates/opsense-rhai/src/ts_ops.rs
@@ -106,7 +106,7 @@ pub fn register(engine: &mut rhai::Engine) {
     engine.register_fn("ts_moving_avg", |points: Array, window_secs: i64| -> Dynamic {
         let Ok(p) = parse_points(&points) else { return Dynamic::from(Array::new()); };
         if p.is_empty() || window_secs <= 0 { return Dynamic::from(Array::new()); }
-        let ws = window_secs as i64;
+        let ws = window_secs;
         let mut out = Array::new();
         for i in 0..p.len() {
             let cutoff = p[i].0 - ws;
@@ -125,7 +125,7 @@ pub fn register(engine: &mut rhai::Engine) {
     engine.register_fn("ts_resample", |points: Array, bucket_secs: i64, agg: &str| -> Dynamic {
         let Ok(p) = parse_points(&points) else { return Dynamic::from(Array::new()); };
         if p.is_empty() || bucket_secs <= 0 { return Dynamic::from(Array::new()); }
-        let bs = bucket_secs as i64;
+        let bs = bucket_secs;
         let t0 = p[0].0;
         let t1 = p.last().unwrap().0;
         let num_buckets = ((t1 - t0) / bs + 1) as usize;
